requirements.txt
atlassian-python-api~=3.12
python-gitlab~=2.9
